+++ packages/experimental/src/parties/party-processor.ts
@@ -1,0 +1,11 @@
+//
+// Copyright 2020 DXOS.org
+//
+
+/**
+ *
+ */
+// TODO(burdon): Implement feed iterator methods.
+export class PartyProcessor {
+
+}