+++ packages/experimental/src/proto/gen/testing.d.ts
@@ -695,11 +695,11 @@ export namespace dxos {
             /** Properties of an ItemMutation. */
             interface IItemMutation {
 
-                /** ItemMutation key */
-                key?: (string|null);
+                /** ItemMutation set */
+                set?: (dxos.echo.testing.ItemMutation.IProperty|null);
 
-                /** ItemMutation value */
-                value?: (string|null);
+                /** ItemMutation append */
+                append?: (dxos.echo.testing.ItemMutation.IProperty|null);
             }
 
             /** Represents an ItemMutation. */
@@ -711,11 +711,11 @@ export namespace dxos {
                  */
                 constructor(properties?: dxos.echo.testing.IItemMutation);
 
-                /** ItemMutation key. */
-                public key: string;
+                /** ItemMutation set. */
+                public set?: (dxos.echo.testing.ItemMutation.IProperty|null);
 
-                /** ItemMutation value. */
-                public value: string;
+                /** ItemMutation append. */
+                public append?: (dxos.echo.testing.ItemMutation.IProperty|null);
 
                 /**
                  * Creates a new ItemMutation instance using the specified properties.
@@ -786,6 +786,105 @@ export namespace dxos {
                  * @returns JSON object
                  */
                 public toJSON(): { [k: string]: any };
+            }
+
+            namespace ItemMutation {
+
+                /** Properties of a Property. */
+                interface IProperty {
+
+                    /** Property key */
+                    key?: (string|null);
+
+                    /** Property value */
+                    value?: (string|null);
+                }
+
+                /** Represents a Property. */
+                class Property implements IProperty {
+
+                    /**
+                     * Constructs a new Property.
+                     * @param [properties] Properties to set
+                     */
+                    constructor(properties?: dxos.echo.testing.ItemMutation.IProperty);
+
+                    /** Property key. */
+                    public key: string;
+
+                    /** Property value. */
+                    public value: string;
+
+                    /**
+                     * Creates a new Property instance using the specified properties.
+                     * @param [properties] Properties to set
+                     * @returns Property instance
+                     */
+                    public static create(properties?: dxos.echo.testing.ItemMutation.IProperty): dxos.echo.testing.ItemMutation.Property;
+
+                    /**
+                     * Encodes the specified Property message. Does not implicitly {@link dxos.echo.testing.ItemMutation.Property.verify|verify} messages.
+                     * @param message Property message or plain object to encode
+                     * @param [writer] Writer to encode to
+                     * @returns Writer
+                     */
+                    public static encode(message: dxos.echo.testing.ItemMutation.IProperty, writer?: $protobuf.Writer): $protobuf.Writer;
+
+                    /**
+                     * Encodes the specified Property message, length delimited. Does not implicitly {@link dxos.echo.testing.ItemMutation.Property.verify|verify} messages.
+                     * @param message Property message or plain object to encode
+                     * @param [writer] Writer to encode to
+                     * @returns Writer
+                     */
+                    public static encodeDelimited(message: dxos.echo.testing.ItemMutation.IProperty, writer?: $protobuf.Writer): $protobuf.Writer;
+
+                    /**
+                     * Decodes a Property message from the specified reader or buffer.
+                     * @param reader Reader or buffer to decode from
+                     * @param [length] Message length if known beforehand
+                     * @returns Property
+                     * @throws {Error} If the payload is not a reader or valid buffer
+                     * @throws {$protobuf.util.ProtocolError} If required fields are missing
+                     */
+                    public static decode(reader: ($protobuf.Reader|Uint8Array), length?: number): dxos.echo.testing.ItemMutation.Property;
+
+                    /**
+                     * Decodes a Property message from the specified reader or buffer, length delimited.
+                     * @param reader Reader or buffer to decode from
+                     * @returns Property
+                     * @throws {Error} If the payload is not a reader or valid buffer
+                     * @throws {$protobuf.util.ProtocolError} If required fields are missing
+                     */
+                    public static decodeDelimited(reader: ($protobuf.Reader|Uint8Array)): dxos.echo.testing.ItemMutation.Property;
+
+                    /**
+                     * Verifies a Property message.
+                     * @param message Plain object to verify
+                     * @returns `null` if valid, otherwise the reason why it is not
+                     */
+                    public static verify(message: { [k: string]: any }): (string|null);
+
+                    /**
+                     * Creates a Property message from a plain object. Also converts values to their respective internal types.
+                     * @param object Plain object
+                     * @returns Property
+                     */
+                    public static fromObject(object: { [k: string]: any }): dxos.echo.testing.ItemMutation.Property;
+
+                    /**
+                     * Creates a plain object from a Property message. Also converts values to other types if specified.
+                     * @param message Property
+                     * @param [options] Conversion options
+                     * @returns Plain object
+                     */
+                    public static toObject(message: dxos.echo.testing.ItemMutation.Property, options?: $protobuf.IConversionOptions): { [k: string]: any };
+
+                    /**
+                     * Converts this Property to JSON.
+                     * @returns JSON object
+                     */
+                    public toJSON(): { [k: string]: any };
+                }
             }
 
             /** Properties of a Timeframe. */